refactoring firm side components for the intake; adding the wills requests to the firm side
